requirements.txt
Flask==2.3.3
openai-whisper==20231117
googletrans==4.0.0rc1
yt-dlp==2023.12.30
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.3
requests>=2.31.0
Werkzeug==2.3.7